libjpeg-turbo, mousemove throttling, repeater autoreconnection, fixed crashes, smaller default server send buffer, queued sends for viewer, more queuing to eliminate excess packets due to disabled nagle algorithm, more misc changes and fixes
git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@552 0bd26501-c230-0410-ab9e-a7946180fe3b

+++ vncviewer/SessionDialog.cpp
@@ -225,7 +225,7 @@ BOOL CALLBACK SessionDialog::SessDlgProc(  HWND hwnd,  UINT uMsg,  WPARAM wParam
 					int numericId = _this->m_port;
 
 					int numberOfHosts = 1;
-					for (int i = 0; i < strlen(display); i++) {
+					for (int i = 0; i < (int)strlen(display); i++) {
 						if (display[i] == ';') {
 							numberOfHosts++;
 						}
